+++ demisto_sdk/commands/common/constants.py
@@ -848,11 +848,11 @@ IOC_OUTPUTS_DICT = {
     'ip': {'IP.Address'},
     'url': {'URL.Data'}
 }
-
-PACK_INITIAL_VERSION = '1.0.0'
-PACK_SUPPORT_OPTIONS = ['xsoar', 'partner', 'developer', 'community']
 XSOAR_SUPPORT = "xsoar"
 XSOAR_AUTHOR = "Cortex XSOAR"
+PACK_INITIAL_VERSION = '1.0.0'
+PACK_SUPPORT_OPTIONS = ['xsoar', 'partner', 'developer', 'community']
+
 XSOAR_SUPPORT_URL = "https://www.paloaltonetworks.com/cortex"
 MARKETPLACE_LIVE_DISCUSSIONS = \
     'https://live.paloaltonetworks.com/t5/cortex-xsoar-discussions/bd-p/Cortex_XSOAR_Discussions'
@@ -940,6 +940,22 @@ FETCH_REQUIRED_PARAMS = [
     }
 ]
 
+MAX_FETCH_PARAM = {
+    'name': 'max_fetch',
+    'required': False,
+    'type': 0,
+    'defaultvalue': '50'
+}
+
+# for reference, the defaultvalue and display are not enforced.
+FIRST_FETCH_PARAM = {
+    'defaultvalue': '7 days',
+    'display': 'First fetch timestamp (<number> <time unit>, e.g., 12 hours, 7 days)',
+    'name': 'first_fetch',
+    'required': False,
+    'type': 0
+}
+
 DOCS_COMMAND_SECTION_REGEX = r'(?:###\s{}).+?(?:(?=(?:\n###\s))|(?=(?:\n##\s))|\Z)'
 # Ignore list for all 'run_all_validations_on_file' method
 ALL_FILES_VALIDATION_IGNORE_WHITELIST = [
@@ -965,3 +981,7 @@ VALIDATED_PACK_ITEM_TYPES = [
     'Classifiers',
     'Layouts'
 ]
+
+FIRST_FETCH = 'first_fetch'
+
+MAX_FETCH = 'max_fetch'