+++ demisto_sdk/commands/common/hook_validations/integration.py
@@ -3,8 +3,11 @@ from demisto_sdk.commands.common.constants import (BANG_COMMAND_NAMES,
                                                    DBOT_SCORES_DICT,
                                                    FEED_REQUIRED_PARAMS,
                                                    FETCH_REQUIRED_PARAMS,
+                                                   FIRST_FETCH,
+                                                   FIRST_FETCH_PARAM,
                                                    INTEGRATION_CATEGORIES,
-                                                   IOC_OUTPUTS_DICT,
+                                                   IOC_OUTPUTS_DICT, MAX_FETCH,
+                                                   MAX_FETCH_PARAM,
                                                    PYTHON_SUBTYPES, TYPE_PWSH)
 from demisto_sdk.commands.common.errors import Errors
 from demisto_sdk.commands.common.hook_validations.content_entity_validator import \
@@ -59,7 +62,8 @@ class IntegrationValidator(ContentEntityValidator):
         return not any(answers)
 
     def is_valid_file(self, validate_rn: bool = True, skip_test_conf: bool = False) -> bool:
-        """Check whether the Integration is valid or not
+        """Check whether the Integration is valid or not according to the LEVEL SUPPORT OPTIONS
+        that depends on the contributor type
 
             Args:
                 validate_rn (bool): Whether to validate release notes (changelog) or not.
@@ -68,6 +72,7 @@ class IntegrationValidator(ContentEntityValidator):
             Returns:
                 bool: True if integration is valid, False otherwise.
         """
+
         answers = [
             super().is_valid_file(validate_rn),
             self.is_valid_subtype(),
@@ -85,11 +90,11 @@ class IntegrationValidator(ContentEntityValidator):
             self.is_valid_pwsh(),
             self.is_valid_image(),
             self.is_valid_description(beta_integration=False),
+            self.is_valid_max_fetch_and_first_fetch(),
         ]
 
         if not skip_test_conf:
             answers.append(self.are_tests_configured())
-
         return all(answers)
 
     def are_tests_configured(self) -> bool:
@@ -665,7 +670,7 @@ class IntegrationValidator(ContentEntityValidator):
         """
         fetch_params_exist = True
         if self.current_file.get('script', {}).get('isfetch') is True:
-            params = [_key for _key in self.current_file.get('configuration', [])]
+            params = [dict.copy(_key) for _key in self.current_file.get('configuration', [])]
             for param in params:
                 if 'defaultvalue' in param:
                     param.pop('defaultvalue')
@@ -675,6 +680,43 @@ class IntegrationValidator(ContentEntityValidator):
                                                                                   yaml.dump(param))
                     if self.handle_error(error_message, error_code, file_path=self.file_path):
                         fetch_params_exist = False
+
+        return fetch_params_exist
+
+    def is_valid_max_fetch_and_first_fetch(self) -> bool:
+        """
+        validate that the max_fetch and first_fetch params exist in the yml and the max_fetch has default value
+        Returns:
+            bool. True if the integration is defined as well False otherwise.
+        """
+        fetch_params_exist = True
+        if self.current_file.get('script', {}).get('isfetch') is True:
+            params = self.current_file.get('configuration', [])
+            first_fetch_param = None
+            max_fetch_param = None
+            for param in params:
+                # the common names for the first_fetch param
+                if param.get('name') == FIRST_FETCH:
+                    first_fetch_param = param
+                elif param.get('name') == MAX_FETCH:
+                    max_fetch_param = param
+
+            if not first_fetch_param:
+                error_message, error_code = Errors.parameter_missing_from_yml_not_community_contributor(
+                    'first_fetch', yaml.dump(FIRST_FETCH_PARAM))
+                if self.handle_error(error_message, error_code, file_path=self.file_path):
+                    fetch_params_exist = False
+
+            if not max_fetch_param:
+                error_message, error_code = Errors.parameter_missing_from_yml_not_community_contributor(
+                    'max_fetch', yaml.dump(MAX_FETCH_PARAM))
+                if self.handle_error(error_message, error_code, file_path=self.file_path):
+                    fetch_params_exist = False
+
+            elif not max_fetch_param.get("defaultvalue"):
+                error_message, error_code = Errors.no_default_value_in_parameter('max_fetch')
+                if self.handle_error(error_message, error_code, file_path=self.file_path):
+                    fetch_params_exist = False
 
         return fetch_params_exist